Allow user to specify whether to use gfm
By default, Github does not render README.md files using Github-flavored
markdown. This causes for a bug in Octodown rendering where images would have
<br> tags where they should not, etc. To fix this, we specify an option which
the user can set: `--[no-]gfm`. This option defaults to setting :gfm to false,
assuming that README.md would be the most common use-case. This can later be a
adapted to auto-detect and set :gfm setting by filename if that is desired.